+++ syndicate/core/groups/meta.py
@@ -49,10 +49,10 @@ USER_LOG = get_user_logger()
 
 @click.group(name=GENERATE_META_GROUP_NAME, cls=AliasedCommandsGroup)
 @return_code_manager
-@click.option('--project-path',
+@click.option('--project-path', '-path',
               cls=MultiWordOption, nargs=1,
-              help="Path to the project folder. Default value: the one "
-                   "from the current config if it exists. "
+              help="Path to the project root directory. Default value: "
+                   "the one from the current config if it exists. "
                    "Otherwise - the current working directory",
               callback=resolve_project_path)
 @click.pass_context