+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [1.18.0] - 2025-06-23
+# [1.18.0] - 2025-06-26
 - Added support for DynamoDB `OnDemandThroughput` limitation
 - Added support for `rds_db_cluster` resource
 - Added support for `rds_db_instance` resource
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed issue related to getting the list of available instance types
 - Changed CLI command and option naming convention from underscores to dashes for consistency. 
 Underscore variants remain supported but are hidden from help output
+- Renamed `bundle-bucket-name` option to `deploy-target-bucket` in the command `syndicate generate project`
+- Added `-path` alias to all `--project-path` options in CLI commands
 - Enhanced the error message when the deploy target bucket is missing.
 
 # [1.17.1] - 2025-03-25